+++ test/controllers/mobile_survey_controller_test.exs
@@ -31,12 +31,13 @@ defmodule Ask.MobileSurveyControllerTest do
 
     respondent = insert(:respondent, survey: survey, respondent_group: group)
     phone_number = respondent.sanitized_phone_number
+    token = Respondent.token(respondent.id)
 
     {:ok, broker} = Broker.start_link
     Broker.poll
 
     assert_receive [:ask, ^test_channel, %Respondent{sanitized_phone_number: ^phone_number}, _, ReplyHelper.simple("Contact", message)]
-    assert message == "Please enter http://app.ask.dev/mobile_survey/#{respondent.id}?token=#{Respondent.token(respondent.id)}"
+    assert message == "Please enter http://app.ask.dev/mobile_survey/#{respondent.id}?token=#{token}"
 
     survey = Repo.get(Survey, survey.id)
     assert survey.state == "running"
@@ -45,7 +46,12 @@ defmodule Ask.MobileSurveyControllerTest do
     assert respondent.state == "active"
     # mobile_survey_send_reply_path
 
-    conn = get conn, mobile_survey_path(conn, :get_step, respondent.id)
+    # Check that get_step without token gives error
+    assert_error_sent :bad_request, fn ->
+      get conn, mobile_survey_path(conn, :get_step, respondent.id)
+    end
+
+    conn = get conn, mobile_survey_path(conn, :get_step, respondent.id, %{token: token})
     json = json_response(conn, 200)
 
     assert %{
@@ -56,7 +62,7 @@ defmodule Ask.MobileSurveyControllerTest do
     } = json["step"]
     assert json["progress"] == 0.0
 
-    conn = get conn, mobile_survey_path(conn, :get_step, respondent.id)
+    conn = get conn, mobile_survey_path(conn, :get_step, respondent.id, %{token: token})
     json = json_response(conn, 200)
     assert %{
       "choices" => [],
@@ -66,7 +72,12 @@ defmodule Ask.MobileSurveyControllerTest do
     } = json["step"]
     assert json["progress"] == 0.0
 
-    conn = post conn, mobile_survey_path(conn, :send_reply, respondent.id, %{value: "", step_id: "s1"})
+    # Check that send_reply without token gives error
+    assert_error_sent :bad_request, fn ->
+      post conn, mobile_survey_path(conn, :send_reply, respondent.id, %{value: "", step_id: "s1"})
+    end
+
+    conn = post conn, mobile_survey_path(conn, :send_reply, respondent.id, %{token: token, value: "", step_id: "s1"})
     json = json_response(conn, 200)
     assert %{
       "choices" => [["Yes"], ["No"]],
@@ -76,7 +87,7 @@ defmodule Ask.MobileSurveyControllerTest do
     } = json["step"]
     assert json["progress"] == 20.0
 
-    conn = get conn, mobile_survey_path(conn, :get_step, respondent.id)
+    conn = get conn, mobile_survey_path(conn, :get_step, respondent.id, %{token: token})
     json = json_response(conn, 200)
     assert %{
       "choices" => [["Yes"], ["No"]],
@@ -86,7 +97,7 @@ defmodule Ask.MobileSurveyControllerTest do
     } = json["step"]
     assert json["progress"] == 20.0
 
-    conn = get conn, mobile_survey_path(conn, :get_step, respondent.id)
+    conn = get conn, mobile_survey_path(conn, :get_step, respondent.id, %{token: token})
     json = json_response(conn, 200)
     assert %{
       "choices" => [["Yes"], ["No"]],
@@ -100,7 +111,7 @@ defmodule Ask.MobileSurveyControllerTest do
     respondent = Repo.get(Respondent, respondent.id)
     assert respondent.disposition == nil
 
-    conn = post conn, mobile_survey_path(conn, :send_reply, respondent.id, %{value: "Yes", step_id: "s2"})
+    conn = post conn, mobile_survey_path(conn, :send_reply, respondent.id, %{token: token, value: "Yes", step_id: "s2"})
     json = json_response(conn, 200)
     assert %{
       "choices" => [["Yes"], ["No"]],
@@ -114,18 +125,19 @@ defmodule Ask.MobileSurveyControllerTest do
     respondent = Repo.get(Respondent, respondent.id)
     assert respondent.disposition == "partial"
 
-    conn = post conn, mobile_survey_path(conn, :send_reply, respondent.id, %{value: "Yes", step_id: "s4"})
+    conn = post conn, mobile_survey_path(conn, :send_reply, respondent.id, %{token: token, value: "Yes", step_id: "s4"})
     json = json_response(conn, 200)
     assert %{
       "prompts" => ["Which is the second perfect number??"],
       "title" => "Which is the second perfect number?",
-      "type" => "numeric"
+      "type" => "numeric",
+      "refusal" => "skip me"
     } = json["step"]
     assert json["progress"] == 60.0
     assert json["error_message"] == "Invalid value"
 
     # Reply a previous step (should reply with the current step)
-    conn = post conn, mobile_survey_path(conn, :send_reply, respondent.id, %{value: "Yes", step_id: "s2"})
+    conn = post conn, mobile_survey_path(conn, :send_reply, respondent.id, %{token: token, value: "Yes", step_id: "s2"})
     json = json_response(conn, 200)
     assert %{
       "prompts" => ["Which is the second perfect number??"],
@@ -135,7 +147,7 @@ defmodule Ask.MobileSurveyControllerTest do
     assert json["progress"] == 60.0
     assert json["error_message"] == "Invalid value"
 
-    conn = post conn, mobile_survey_path(conn, :send_reply, respondent.id, %{value: "99", step_id: "s5"})
+    conn = post conn, mobile_survey_path(conn, :send_reply, respondent.id, %{token: token, value: "99", step_id: "s5"})
     json = json_response(conn, 200)
     assert %{
       "prompts" => ["What's the number of this question??"],
@@ -144,7 +156,7 @@ defmodule Ask.MobileSurveyControllerTest do
     } = json["step"]
     assert json["progress"] == 80.0
 
-    conn = post conn, mobile_survey_path(conn, :send_reply, respondent.id, %{value: "11", step_id: "s6"})
+    conn = post conn, mobile_survey_path(conn, :send_reply, respondent.id, %{token: token, value: "11", step_id: "s6"})
     json = json_response(conn, 200)
     assert %{
       "prompts" => ["The survey has ended"],
@@ -206,6 +218,7 @@ defmodule Ask.MobileSurveyControllerTest do
     RespondentGroupChannel.changeset(%RespondentGroupChannel{}, %{respondent_group_id: group.id, channel_id: channel.id, mode: "mobileweb"}) |> Repo.insert
 
     respondent = insert(:respondent, survey: survey, respondent_group: group)
+    token = Respondent.token(respondent.id)
 
     {:ok, _} = Broker.start_link
     Broker.poll
@@ -213,7 +226,7 @@ defmodule Ask.MobileSurveyControllerTest do
     respondent = Repo.get(Respondent, respondent.id)
     respondent |> Respondent.changeset(%{"state" => "finised"}) |> Repo.update!
 
-    conn = get conn, mobile_survey_path(conn, :get_step, respondent.id)
+    conn = get conn, mobile_survey_path(conn, :get_step, respondent.id, %{token: token})
     assert %{
       "prompts" => ["The survey has ended"],
       "title" => "The survey has ended",
@@ -232,13 +245,14 @@ defmodule Ask.MobileSurveyControllerTest do
     RespondentGroupChannel.changeset(%RespondentGroupChannel{}, %{respondent_group_id: group.id, channel_id: channel.id, mode: "mobileweb"}) |> Repo.insert
 
     respondent = insert(:respondent, survey: survey, respondent_group: group)
+    token = Respondent.token(respondent.id)
 
     {:ok, _} = Broker.start_link
     Broker.poll
 
     survey |> Survey.changeset(%{"state" => "completed"}) |> Repo.update!
 
-    conn = get conn, mobile_survey_path(conn, :get_step, respondent.id)
+    conn = get conn, mobile_survey_path(conn, :get_step, respondent.id, %{token: token})
     assert %{
       "prompts" => ["Bye"],
       "title" => "Bye",
@@ -289,5 +303,51 @@ defmodule Ask.MobileSurveyControllerTest do
 
     conn = get conn, mobile_survey_path(conn, :index, respondent.id, %{token: token})
     assert response(conn, 200)
+  end
+
+  test "respondent flow via mobileweb with refusal + end", %{conn: conn} do
+    test_channel = TestChannel.new(false, true)
+
+    channel = insert(:channel, settings: test_channel |> TestChannel.settings, type: "sms")
+    quiz = insert(:questionnaire, steps: @mobileweb_refusal_dummy_steps, settings: %{"error_message" => %{"en" => %{"mobileweb" => "Invalid value"}}})
+    survey = insert(:survey, Map.merge(@always_schedule, %{state: "running", questionnaires: [quiz], mode: [["mobileweb"]]}))
+    group = insert(:respondent_group, survey: survey, respondents_count: 1) |> Repo.preload(:channels)
+
+    RespondentGroupChannel.changeset(%RespondentGroupChannel{}, %{respondent_group_id: group.id, channel_id: channel.id, mode: "mobileweb"}) |> Repo.insert
+
+    respondent = insert(:respondent, survey: survey, respondent_group: group)
+    phone_number = respondent.sanitized_phone_number
+    token = Respondent.token(respondent.id)
+
+    {:ok, broker} = Broker.start_link
+    Broker.poll
+
+    assert_receive [:ask, ^test_channel, %Respondent{sanitized_phone_number: ^phone_number}, _, ReplyHelper.simple("Contact", message)]
+    assert message == "Please enter http://app.ask.dev/mobile_survey/#{respondent.id}?token=#{token}"
+
+    survey = Repo.get(Survey, survey.id)
+    assert survey.state == "running"
+
+    respondent = Repo.get(Respondent, respondent.id)
+    assert respondent.state == "active"
+    # mobile_survey_send_reply_path
+
+    conn = get conn, mobile_survey_path(conn, :get_step, respondent.id, %{token: token})
+    json = json_response(conn, 200)
+
+    assert %{
+      "choices" => [],
+      "prompts" => ["Welcome to the survey!"],
+      "title" => "Let there be rock",
+      "type" => "explanation"
+    } = json["step"]
+    assert json["progress"] == 0.0
+
+    post conn, mobile_survey_path(conn, :send_reply, respondent.id, %{token: token, value: "", step_id: "s1"})
+
+    respondent = Repo.get(Respondent, respondent.id)
+    assert respondent.disposition == "refused"
+
+    :ok = broker |> GenServer.stop
   end
 end