+++ test/support/dummy_steps.ex
@@ -762,7 +762,15 @@ defmodule Ask.DummySteps do
             ),
           store: "Perfect Number",
           skip_logic: default_numeric_skip_logic(),
-          refusal: nil
+          refusal: %{
+            "enabled" => true,
+            "responses" => %{
+              "mobileweb" => %{
+                "en" => "skip me",
+              }
+            },
+            "skip_logic" => "end",
+          }
         ),
         numeric_step(
           id: "s6",
@@ -774,6 +782,23 @@ defmodule Ask.DummySteps do
           skip_logic: default_numeric_skip_logic(),
           refusal: nil
         )
+      ]
+
+      @mobileweb_refusal_dummy_steps [
+        explanation_step(
+          id: "s1",
+          title: "Let there be rock",
+          prompt: prompt(
+            mobileweb: sms_prompt("Welcome to the survey!")
+          ),
+          skip_logic: nil
+        ),
+        flag_step(
+          id: "s3",
+          title: "Let there be rock",
+          disposition: "refused",
+          skip_logic: "end"
+        ),
       ]
     end
   end